scarf/requirements.txt
scipy
streamlit
pillow
numpy
